chore: update Dockerfile and GitHub Actions workflow to support configurable binary name
The Dockerfile has been updated to use the `BINARY_NAME` build argument to specify the name of the binary to be copied into the image. The `ENTRYPOINT` has also been updated to use the `BINARY_NAME` environment variable.

The GitHub Actions workflow has been updated to pass the `BINARY_NAME` environment variable as a build argument to the Docker build step.